Refactor out GameType
I put it in a utils package since it's used in the controller, model,
and view.
I might need to define the game types themselves in the controllers
package, since they depend on both the model and the view.
We should also follow the java package naming convention and put our
folders under `me.quickga._`, since now we have to import in the Build
file under `_root_`, which is unweildy.
I want to also do some refactoring and take out the view's dependencies
on the game type.